general gameplay works now
i tried refactoring some of my code so i can solve the issue with
the end game function

it seems like when i define player names/markers in the game, it
is not looking at the players object... but i'm not sure then
how it still manages to define those player names/markers?

i tried playersData.players={} in end game function to override the
existing players object, and when i console log the object, it
shows the new player names/markers, but when actually playing
the game, the names/markers are the same as the initial inputs

also added more comments so i can try to figure out the scopes..

i was suggested to remove the bind for MakeMove in the start game
event listener, so will try that next.

as for CSS, i don't remember how to make a square grid that resizes
with the window...